add context check

+++ app/graphql/mutations/delete_user.rb
@@ -4,6 +4,8 @@ module Mutations
     field :user, Types::UserType, null: false
 
     def resolve
+      return unless context[:current_user]
+
       user = context[:current_user]
       user.destroy!
